+++ packages/datepicker/CHANGELOG.md
@@ -1,6 +1,22 @@
 # @heathmont/moon-datepicker
 
+## 8.1.0
+
+### Minor Changes
+
+- TextInput Groups
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@8.1.0
+  - @heathmont/moon-themes@8.1.0
+  - @heathmont/moon-components@8.1.0
+  - @heathmont/moon-icons@8.1.0
+  - @heathmont/moon-utils@8.1.0
+
 ## 8.0.0
+
 ### Major Changes
 
 - New tokens