+++ packages/draggabletable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 8.1.0
+
+### Minor Changes
+
+- TextInput Groups
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@8.1.0
+  - @heathmont/moon-components@8.1.0
+  - @heathmont/moon-table@8.1.0
+  - @heathmont/moon-utils@8.1.0
+
 ## 8.0.0
 
 ### Major Changes